Move effects.py to scenes package.
