+++ src/models/scenes/effects_test.py
@@ -2,11 +2,10 @@ import pytest
 
 from models.characters.character_examples import CharacterTypes
 from models.characters.character_impl import build_character
-from models.characters.effects import (AcquireMod, IncrementAttribute,
-                                       RestartGame)
 from models.characters.mods_base import build_mod
 from models.characters.player import get_player, reset_player
 from models.characters.states import Attributes
+from models.scenes.effects import AcquireMod, IncrementAttribute, RestartGame
 
 
 @pytest.fixture(scope='function')
