+++ src/models/scenes/scene_examples.py
@@ -4,7 +4,6 @@ from functools import partial
 from typing import Dict, Tuple, cast
 
 from data.constants import BackgroundImages
-from models.characters.effects import IncrementAttribute, RestartGame
 from models.characters.mods_base import Mod, SlotTypes, build_mod
 from models.characters.player import get_player
 from models.characters.states import Attributes, Skill
@@ -12,6 +11,7 @@ from models.characters.subroutine_examples import direct_damage
 from models.scenes import combat_scene
 from models.scenes.decision_scene import (DecisionOption, DecisionScene,
                                           from_transition, transition_to)
+from models.scenes.effects import IncrementAttribute, RestartGame
 from models.scenes.inventory_scene import InventoryScene
 from models.scenes.scenes_base import BasicResolution, Resolution, Scene
 from models.scenes.skill_checks import Difficulty, skill_check
